backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.3.0